+++ src/Ocelot.Provider.Kubernetes/Interfaces/IEndPointClient.cs
@@ -1,7 +1,7 @@
 ﻿using KubeClient.Models;
 using KubeClient.ResourceClients;
 
-namespace Ocelot.Provider.Kubernetes;
+namespace Ocelot.Provider.Kubernetes.Interfaces;
 
 public interface IEndPointClient : IKubeResourceClient
 {
