+++ src/Ocelot.Provider.Kubernetes/Interfaces/IKubeServiceBuilder.cs
@@ -1,0 +1,9 @@
+﻿using KubeClient.Models;
+using Ocelot.Values;
+
+namespace Ocelot.Provider.Kubernetes.Interfaces;
+
+public interface IKubeServiceBuilder
+{
+    IEnumerable<Service> BuildServices(KubeRegistryConfiguration configuration, EndpointsV1 endpoint);
+}
