+++ src/Ocelot.Provider.Kubernetes/KubernetesProviderFactory.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Ocelot.Configuration;
 using Ocelot.Logging;
+using Ocelot.Provider.Kubernetes.Interfaces;
 
 namespace Ocelot.Provider.Kubernetes
 {
@@ -17,14 +18,16 @@ namespace Ocelot.Provider.Kubernetes
         {
             var factory = provider.GetService<IOcelotLoggerFactory>();
             var kubeClient = provider.GetService<IKubeApiClient>();
+            var serviceBuilder = provider.GetService<IKubeServiceBuilder>();
 
             var configuration = new KubeRegistryConfiguration
             {
                 KeyOfServiceInK8s = route.ServiceName,
                 KubeNamespace = string.IsNullOrEmpty(route.ServiceNamespace) ? config.Namespace : route.ServiceNamespace,
+                Scheme = route.DownstreamScheme,
             };
 
-            var defaultK8sProvider = new Kube(configuration, factory, kubeClient);
+            var defaultK8sProvider = new Kube(configuration, factory, kubeClient, serviceBuilder);
  
             return PollKube.Equals(config.Type, StringComparison.OrdinalIgnoreCase)
                 ? new PollKube(config.PollingInterval, factory, defaultK8sProvider)
