+++ src/Ocelot.Provider.Kubernetes/OcelotBuilderExtensions.cs
@@ -1,16 +1,18 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Ocelot.DependencyInjection;
+using Ocelot.Provider.Kubernetes.Interfaces;
 
-namespace Ocelot.Provider.Kubernetes
+namespace Ocelot.Provider.Kubernetes;
+
+public static class OcelotBuilderExtensions
 {
-    public static class OcelotBuilderExtensions
+    public static IOcelotBuilder AddKubernetes(this IOcelotBuilder builder, bool usePodServiceAccount = true)
     {
-        public static IOcelotBuilder AddKubernetes(this IOcelotBuilder builder, bool usePodServiceAccount = true)
-        {
-            builder.Services
-                .AddSingleton(KubernetesProviderFactory.Get)
-                .AddKubeClient(usePodServiceAccount);
-            return builder;
-        }
+        builder.Services
+            .AddKubeClient(usePodServiceAccount)
+            .AddSingleton(KubernetesProviderFactory.Get)
+            .AddSingleton<IKubeServiceBuilder, KubeServiceBuilder>()
+            .AddSingleton<IKubeServiceCreator, KubeServiceCreator>();
+        return builder;
     }
 }
