+++ src/Ocelot/ServiceDiscovery/ServiceDiscoveryFinderDelegate.cs
@@ -1,7 +1,6 @@
 using Ocelot.Configuration;
 using Ocelot.ServiceDiscovery.Providers;
 
-namespace Ocelot.ServiceDiscovery
-{
-    public delegate IServiceDiscoveryProvider ServiceDiscoveryFinderDelegate(IServiceProvider provider, ServiceProviderConfiguration config, DownstreamRoute route);
-}
+namespace Ocelot.ServiceDiscovery;
+
+public delegate IServiceDiscoveryProvider ServiceDiscoveryFinderDelegate(IServiceProvider provider, ServiceProviderConfiguration config, DownstreamRoute route);
