+++ test/Ocelot.UnitTests/Kubernetes/KubeTests.cs
@@ -6,11 +6,12 @@ using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 using Ocelot.Logging;
 using Ocelot.Provider.Kubernetes;
+using Ocelot.Provider.Kubernetes.Interfaces;
 using Ocelot.Values;
 
 namespace Ocelot.UnitTests.Kubernetes
 {
-    public class KubeTests : UnitTest, IDisposable
+    public class KubeTests : IDisposable
     {
         private IWebHost _fakeKubeBuilder;
         private readonly Kube _provider;
@@ -21,10 +22,11 @@ namespace Ocelot.UnitTests.Kubernetes
         private readonly string _kubeHost;
         private readonly string _fakekubeServiceDiscoveryUrl;
         private List<Service> _services;
+        private string _receivedToken;
         private readonly Mock<IOcelotLoggerFactory> _factory;
         private readonly Mock<IOcelotLogger> _logger;
-        private string _receivedToken;
         private readonly IKubeApiClient _clientFactory;
+        private readonly Mock<IKubeServiceBuilder> _serviceBuilder;
 
         public KubeTests()
         {
@@ -33,8 +35,8 @@ namespace Ocelot.UnitTests.Kubernetes
             _port = 5567;
             _kubeHost = "localhost";
             _fakekubeServiceDiscoveryUrl = $"{Uri.UriSchemeHttp}://{_kubeHost}:{_port}";
-            _endpointEntries = new EndpointsV1();
-            _factory = new Mock<IOcelotLoggerFactory>();
+            _endpointEntries = new();
+            _factory = new();
 
             var option = new KubeClientOptions
             {
@@ -45,19 +47,21 @@ namespace Ocelot.UnitTests.Kubernetes
             };
 
             _clientFactory = KubeApiClient.Create(option);
-            _logger = new Mock<IOcelotLogger>();
+            _logger = new();
             _factory.Setup(x => x.CreateLogger<Kube>()).Returns(_logger.Object);
             var config = new KubeRegistryConfiguration
             {
                 KeyOfServiceInK8s = _serviceName,
                 KubeNamespace = _namespaces,
             };
-            _provider = new Kube(config, _factory.Object, _clientFactory);
+            _serviceBuilder = new();
+            _provider = new Kube(config, _factory.Object, _clientFactory, _serviceBuilder.Object);
         }
 
         [Fact]
         public void Should_return_service_from_k8s()
         {
+            // Arrange
             var token = "Bearer txpc696iUhbVoudg164r93CxDTrKRVWG";
             var endPointEntryOne = new EndpointsV1
             {
@@ -65,6 +69,7 @@ namespace Ocelot.UnitTests.Kubernetes
                 ApiVersion = "1.0",
                 Metadata = new ObjectMetaV1
                 {
+                    Name = nameof(Should_return_service_from_k8s),
                     Namespace = "dev",
                 },
             };
@@ -79,13 +84,17 @@ namespace Ocelot.UnitTests.Kubernetes
                 Port = 80,
             });
             endPointEntryOne.Subsets.Add(endpointSubsetV1);
+            _serviceBuilder.Setup(x => x.BuildServices(It.IsAny<KubeRegistryConfiguration>(), It.IsAny<EndpointsV1>()))
+                .Returns(new Service[] { new(nameof(Should_return_service_from_k8s), new("localhost", 80), string.Empty, string.Empty, new string[0]) });
+            GivenThereIsAFakeKubeServiceDiscoveryProvider(_fakekubeServiceDiscoveryUrl, _serviceName, _namespaces);
+            GivenTheServicesAreRegisteredWithKube(endPointEntryOne);
 
-            this.Given(x => GivenThereIsAFakeKubeServiceDiscoveryProvider(_fakekubeServiceDiscoveryUrl, _serviceName, _namespaces))
-                .And(x => GivenTheServicesAreRegisteredWithKube(endPointEntryOne))
-                .When(x => WhenIGetTheServices())
-                .Then(x => ThenTheCountIs(1))
-                .And(_ => ThenTheTokenIs(token))
-                .BDDfy();
+            // Act
+            WhenIGetTheServices();
+
+            // Assert
+            ThenTheCountIs(1);
+            ThenTheTokenIs(token);
         }
 
         private void ThenTheTokenIs(string token)
